Resolve "As an admin I would like to add official tags onto official activities"

Closes #24

See merge request c1936922/group-03-staff-project!38

+++ src/main/java/group03/project/domain/Reflection.java
@@ -15,11 +15,11 @@ public class Reflection {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="reflectionID")
-    private Integer reflectionID;
+    private Long reflectionID;
     @Column(name="Participation_participationID")
-    private Integer participationID;
+    private Long participationID;
     @Column(name="Tag_tagID")
-    private Integer tagID;
+    private Long tagID;
     @Column(name="Reflect_what")
     private String reflect_what;
     @Column(name="Reflect_prompt")
